fixing test cases
